RaspberryPiCode/requirements.txt
python-chess
pyserial
requests
Pillow
